requirements.txt
PyYAML
black
boto3
